Move SimReady search resolver to utils.assets
The search function is spawner-agnostic: it takes a query and returns
USD asset paths. Move it and the endpoint constant to
isaaclab.utils.assets, next to the other asset retrieval helpers, so
non-spawner consumers can use it without importing spawner modules.
The simready spawner sub-module keeps only the two cfg subclasses.

Split the tests to mirror the layout: resolver tests move to the
utils asset tests together with the fake simready.search module; the
spawner cfg tests stub the resolver directly and gain a
parameter-forwarding test.
